refactor(CO-3211): update filters type imports to use specific paths for better clarity
