+++ src/types/index.d.ts
@@ -10,7 +10,6 @@ export * from './api';
 export * from './attachments';
 export * from './backup-search';
 export * from './conversations';
-export * from './filters';
 export * from './messages';
 export * from './participant';
 export * from './soap';
