+++ src/views/settings/filters/parts/create-filter-modal.tsx
@@ -12,7 +12,7 @@ import { map, omit, reduce } from 'lodash';
 import { useTranslation } from 'react-i18next';
 import { v4 as uuidv4 } from 'uuid';
 
-import type { Filter, FilterActions } from 'types/index.d';
+import { Filter, FilterActions } from 'types/filters';
 import { CreateFilterContext } from 'views/settings/filters/parts/create-filter-context';
 import ModalFooter from 'views/settings/filters/parts/create-filter-modal-footer';
 import DefaultCondition from 'views/settings/filters/parts/create-filters-conditions/default';
