+++ src/views/settings/filters/parts/filter-actions-panel.tsx
@@ -9,7 +9,7 @@ import { Container, Text } from '@zextras/carbonio-design-system';
 import { map } from 'lodash';
 import { useTranslation } from 'react-i18next';
 
-import { FilterAction } from 'types/index.d';
+import { FilterAction } from 'types/filters';
 import Heading from 'views/settings/components/settings-heading';
 import { FilterActionRow } from 'views/settings/filters/parts/filter-action-row';
 import { getActionTranslations } from 'views/settings/filters/parts/utils';
