+++ src/views/settings/filters/parts/tests/filter-actions-panel.test.tsx
@@ -9,8 +9,10 @@ import React from 'react';
 import { screen, within } from '@testing-library/react';
 
 import { setupTest } from '@test-setup';
-import { FilterActionsProps } from 'types/index.d';
-import { FilterActionsPanel } from 'views/settings/filters/parts/filter-actions-panel';
+import {
+	FilterActionsPanel,
+	FilterActionsProps
+} from 'views/settings/filters/parts/filter-actions-panel';
 
 describe('FilterActionsPanel', () => {
 	it('should update actions when switching an existing action for another one', async () => {
