+++ src/views/settings/filters/parts/tests/modify-filter-modal.test.tsx
@@ -10,7 +10,7 @@ import React, { act } from 'react';
 import { screen } from '@testing-library/react';
 
 import { setupTest } from '@test-setup';
-import { Filter } from 'types/index.d';
+import { Filter } from 'types/filters';
 import { ModifyFilterModal } from 'views/settings/filters/parts/modify-filter/modify-filter-modal';
 
 describe('modify filter modal', () => {
